// ui/lib/api/schemas/engineCatalog.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 */
import type { EngineCatalogEntry } from './engineCatalogEntry'

export interface EngineCatalog {
  bubbleSegmenters: EngineCatalogEntry[]
  detectors: EngineCatalogEntry[]
  fontDetectors: EngineCatalogEntry[]
  inpainters: EngineCatalogEntry[]
  ocr: EngineCatalogEntry[]
  renderers: EngineCatalogEntry[]
  segmenters: EngineCatalogEntry[]
  translators: EngineCatalogEntry[]
}